requirements/development.txt
-r base.txt
django-shortcuts==1.6
